add serde module and lz4 to benches
