+++ execution-plane/arcon/src/streaming/channel/mod.rs
@@ -1,13 +1,7 @@
-use crate::data::{reliable_remote::ReliableSerde, unsafe_remote::UnsafeSerde};
-use crate::data::{ArconMessage, ArconType};
-use kompact::prelude::{ActorPath, ActorRefStrong};
 pub mod strategy;
 
-#[derive(Clone)]
-pub enum ArconSerde<A: ArconType> {
-    Unsafe(UnsafeSerde<A>),
-    Reliable(ReliableSerde<A>),
-}
+use crate::data::{serde::ArconSerde, ArconMessage, ArconType};
+use kompact::prelude::{ActorPath, ActorRefStrong};
 
 #[derive(Clone)]
 pub enum Channel<A: ArconType> {
